page load for meals is broken out
